backend/requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2 ; python_version >= '3.7'
coverage[toml]==7.3.1 ; python_version >= '3.8'
django==4.2.5
djangorestframework==3.14.0
iniconfig==2.0.0 ; python_version >= '3.7'
packaging==23.1 ; python_version >= '3.7'
pillow==10.0.1
pluggy==1.3.0 ; python_version >= '3.8'
pytest==7.4.2
pytest-cov==4.1.0
pytest-django==4.5.2
python-magic==0.4.27
pytz==2023.3.post1
sqlparse==0.4.4 ; python_version >= '3.5'
